Skip header data when 0xFE81 interlacing indicator is found
The 0xFE81 hexadecimal seems to indicate that an image is interlaced. This signature is typically followed by data that doesn't seem to get rendered into the image.

It's currently not understood how to figure out how much data should get skipped. For many images this is just 8, but for the load00.rle it seems to be 18.

For now this is added to document and start the initial behavior for accounting for this. It will be developed further down the line.