Fix Flag/Quantity reading to use bits from Flag byte in quantity
Originally it was assumed that an entire byte was used to determine the flag/functionality along with 1 byte for the quantity.

This was inaccurate. The quantity can actually be larger than a byte, and it uses 15 bits for the quantity from the 2 bytes, and the flag portion is only a single bit. This was discovered by research done by DKDave.

The most significant bit is a toggle bit that controls the repeat color or read X colors functionality.