Remove Newline functionality causing inaccurate reads
The Newline flag was a misinterpretation of the flag flag variations. While it worked accurately for some images, this was coincidental and caused incorrect reading behavior for other images which were perceived as edge cases of more complex functionality.

Edge cases with 0xFE 81 were also resulting from a misunderstanding of the quantity amounts.

Wrapping functionality can be removed completely as the images that seemed to use it were always still 512 in width.

This simplifies the logic notably.