Update to not include extraneous black row
The image is intended to be 512x240, however the image contains an extra black row in the data. This is not meant to be displayed, and we can instead stop reading once we've hit the total rows needed.

The number of rows can be determined by getting the decompressed header (4 byte integer that follows the 8 byte magic number), subtracting the size of the magic number from it, and dividing it by the width of the image (a fixed 512).

We simply stop reading once we've added the final row.