Display GIMP message directly & not suppress dialog with exception
Previously the use of the alert_and_raise function caused messages to not get bubbled up to the user to display the errors/issues due to the exception that would get raised immediately after the pdb.gimp_message causing the plugin to exit before finishing the message.

If this method only does two actions, and one is removed, we should simply display the image directly and return to stop the plugin.