Adds functionality to automatically detect and fix columns shifted to the opposite end
For RLE images, there are often two rows that should be on the left side of the image that are actually written as the final two columns. When this happens, it can be identified by the top two pixels of the being red (on the left) and blue (on the right).

This code identifies it from the blue shade and if it is detected, swaps the columns to the start of the image, and then shifts the pixels from each row of those columns up to replace the red/blue pixels that mark them. (Without this, the rows would be offset one too low).